BlockPrecisionMatrix.jl

A Visual Explanation of Statistical Testing

In the animation above, they illustrate a test on the mean. In our case, we want to apply the ideas of permutation tests (inverse of the covariance matrix) but to identify the null blocks of a precision matrix. A block will be null in the following situation.

Consider the blocks associated with the variables $X$ and $Y$ of a data set D such that

$D = (X,Y,Z)$ with $X \in R^{n,p_X}, Y \in R^{n,p_Y}, Z \in R^{n,p_Z}$

\[\begin{aligned} X & = \beta_X [1,Z] + \epsilon_X \\ Y & = \beta_Y [1,Z] + \epsilon_Y \end{aligned}\]

then the block $P_{XY}$ is null if $\epsilon_X$ and $\epsilon_Y$ are independent ie if the correlation matrix of $\epsilon_X$ with $\epsilon_Y$ is null.

So we apply the permutations to the residues $\epsilon_X$ and/or $\epsilon_Y$. To do this, we first need to estimate the linear model (this is where we use the SCAD estimation), recover the residuals, permute them and then reconstruct the data corresponding to the permuted residuals with permute_scad.

To finish the test, we need to estimate the precision matrices associated with the permutations. We use the callable type StatTest.

BlockPrecisionMatrix.permute_scad — Function

permute_scad(rng, Y, Z)

Reconstruct the data corresponding to the permuted residuals esimated with SCAD regression.

BlockPrecisionMatrix.StatTest — Type

StatTest(n, p)

We use the Xia estimator for the precision matrix

BlockPrecisionMatrix.cov_simu — Function

cov_simu(rng, blocs, indblocs, blocs_on, D)

fonction pour simuler des matrices de covariances avec blocs

blocs : indice des intervalles separant les blocs (sortie de StructureCov)

indblocs : liste des indices des blocs (sortie de StructureCov)

blocs_on : liste des indices des couples de blocs "allumes"

D : vecteur des valeurs propres de la matrice a simuler

BlockPrecisionMatrix.structure_cov — Function

structure_cov(rng, p, b, blocs_on)

Fonction pour definir les indices des blocs

p : dimension des donnees

b : nombre de blocs

blocs_on : liste des indices des couples de blocs "allumes"

BlockPrecisionMatrix.generate_data — Function

generate_data(rng, p, n, b, blocs_on )

Create the simulation data and returning the covariance matrix and prediction matrix with blocks.

Example

using Random
using UnicodePlots
import BlockPrecisionMatrix.generate_data

p = 10 
n = 500
b = 3
rng = MersenneTwister(12)
blocs_on  = [[1,3]]

covmat, premat, data, blocks = generate_data(rng, p, n, b, blocs_on)

display(heatmap(covmat, title="covmat"))
display(heatmap(premat, title="premat"))

BlockPrecisionMatrix.permute — Function

permute(rng, x, n)

Permutation

BlockPrecisionMatrix.iwt_block_precision — Function

iwt_block_precision(data, blocks; B=1000, estimation=:SCAD)

function performing the test on blocks and adjusting the results

data: data matrix (n*p) where n is the sample size and p the number of grid points (for now. it will be spline coefficients?)

blocks: vector of length p containing block indexes (numeric, from 1 to number of blocks) for each grid point

B: number of permutations done to evaluate the tests p-values